Module Irmin_watcher__.Core

module Digests : sig ... end

Sets of filenames and their digests.

module Dispatch : sig ... end

Dispatch listening functions.

module Watchdog : sig ... end

Watchdog functions. Ensure that only one background process is monitoring events for a given directory.

type hook = int -> string -> (string -> unit Lwt.t) -> (unit -> unit Lwt.t) Lwt.t

The type for Irmin.Watch hooks.

type t

The type for listeners.

val create : (string -> Watchdog.hook) -> t

create h is the Irmin watcher using the update hook h.

val watchdog : t -> Watchdog.t

watchdog t is t's watchdog.

val hook : t -> hook

hook t is t's hook.

Helpers

val stoppable : (unit -> unit Lwt.t) -> unit -> unit Lwt.t

stoppable t is a function f such that calling f will cancel the thread t.

val default_polling_time : float Stdlib.ref