Module Irmin_watcher__Backend

FSevents backend for Irmin watchers.

0.5.0 — homepage

val v : Irmin_watcher__.Core.t Stdlib.Lazy.t

v id p f is the hook calling f everytime a sub-path of p is modified. Return a function to call to remove the hook. Use the FSevent framework to be notified on filesystem changes.

val mode : [ `FSEvents ]