Module Irmin_watcher

Irmin watchers.

0.3.0 — homepage

val v : Irmin_watcher__Core.t

v id p f is the listen hook calling f everytime a sub-path of p is modified. Return a function to call to remove the hook. Default to polling if no better solution is available. FSevents and Inotify backends are available.

val mode : [ `FSEvents | `Inotify | `Polling ]

type stats = {

watchdogs : int;

dispatches : int;

}

val hook : int -> string -> (string -> unit Lwt.t) -> (unit -> unit Lwt.t) Lwt.t

hook t is an Irmin.Watcher compatible representation of v.

val stats : unit -> stats

stats () is a snapshot of v's stats.

val set_polling_time : float -> unit

set_polling_time f set the polling interval to f. Only valid when mode = `Polling.